refactor: PaymentExceptionHandler BC 경계 분리 및 CoreFacade 미사용 메서드 제거
