+++ bc/core/src/main/java/app/giftify/payment/adapter/inbound/web/exception/PaymentExceptionHandler.java
@@ -1,24 +1,16 @@
 package app.giftify.payment.adapter.inbound.web.exception;
 
 import app.giftify.payment.domain.PaymentException;
-import app.giftify.wallet.domain.WalletException;
 import app.giftify.shared.api.response.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@RestControllerAdvice(basePackages = {"app.giftify.payment", "app.giftify.wallet"})
+@RestControllerAdvice(basePackages = {"app.giftify.payment"})
 public class PaymentExceptionHandler {
 
 	@ExceptionHandler(PaymentException.class)
 	public ResponseEntity<ErrorResponse> handlePaymentException(PaymentException e) {
-		return ResponseEntity
-				.status(e.getErrorCode().getStatusCode())
-				.body(new ErrorResponse(e.getErrorCode().getCode(), e.getMessage()));
-	}
-
-	@ExceptionHandler(WalletException.class)
-	public ResponseEntity<ErrorResponse> handleWalletException(WalletException e) {
 		return ResponseEntity
 				.status(e.getErrorCode().getStatusCode())
 				.body(new ErrorResponse(e.getErrorCode().getCode(), e.getMessage()));
